backend/requirements.txt
# Core LangChain packages
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langgraph>=0.3.0,<0.4.0
langgraph-supervisor>=0.0.1,<0.1.0

# LangGraph BigTool for advanced tool management
langgraph-bigtool>=0.0.3,<0.1.0

# Google AI integration
langchain-google-genai>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0

# OpenAI for embeddings (used by BigTool)
openai>=1.0.0,<2.0.0
langchain-openai>=0.2.0,<0.4.0

# MongoDB vector store dependencies
pymongo>=4.0.0,<5.0.0
motor>=3.0.0,<4.0.0  # Async MongoDB driver for potential async support

# GUI
customtkinter>=5.0.0

# Core utilities
pydantic>=2.0.0,<3.0.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0
google-cloud-texttospeech>=2.0.0,<3.0.0
google-cloud-speech>=2.0.0,<3.0.0

# Image processing
google-cloud-vision>=3.0.0,<4.0.0
pillow>=9.0.0,<11.0.0

# Additional dependencies
requests>=2.25.0
typing-extensions>=4.0.0
jsonschema>=4.0.0
tenacity>=8.0.0 